+++ addons/yat/src/commands/Cls.cs
@@ -7,8 +7,6 @@ namespace YAT.Commands;
 [Command("cls", "Clears the console.", aliases: "clear")]
 public sealed class Cls : ICommand
 {
-    public string[]? Arguments { get; set; }
-
     public CommandResult Execute(CommandData data)
     {
         data.Terminal.Clear();