+++ addons/yat/src/commands/Crash.cs
@@ -9,13 +9,9 @@ namespace YAT.Commands;
 [Option("-message", "string", "The message to when crashing.", ":c")]
 public sealed class Crash : ICommand
 {
-    public string[]? Arguments { get; set; }
-
     public CommandResult Execute(CommandData data)
     {
-        var message = (string)data.Options["-message"];
-
-        OS.Crash(message);
+        OS.Crash((string)data.Options["-message"]);
 
         return ICommand.Success();
     }