+++ addons/yat/src/commands/Forcegc.cs
@@ -18,8 +18,6 @@ namespace YAT.Commands;
 [Option("-c", "bool", "A value that specifies whether the GC is forced to compact the heap.")]
 public sealed class Forcegc : ICommand
 {
-    public string[]? Arguments { get; set; }
-
     public CommandResult Execute(CommandData data)
     {
         var gen = (int)data.Options["-g"];