+++ addons/yat/src/commands/Sys.cs
@@ -16,8 +16,6 @@ namespace YAT.Commands;
 [Threaded]
 public sealed class Sys : ICommand
 {
-    public string[]? Arguments { get; set; }
-
     public CommandResult Execute(CommandData data)
     {
         var program = (string)data.Options["-program"];