+++ addons/yat/src/commands/View.cs
@@ -18,13 +18,11 @@ namespace YAT.Commands;
 public sealed class View : ICommand
 {
     private static readonly int MAX_DRAW_MODE;
-    private static readonly List<StringName> Modes = new();
-
-    public string[]? Arguments { get; set; }
+    private static readonly List<StringName> Modes = [];
 
     static View()
     {
-        Array values = Enum.GetValues(typeof(ViewportDebugDraw));
+        Array values = Enum.GetValues<ViewportDebugDraw>();
         MAX_DRAW_MODE = values.Length - 1;
 
         foreach (object? mode in values)
@@ -46,7 +44,7 @@ public sealed class View : ICommand
         if (Modes.Contains(mode))
         {
             return SetDebugDraw(
-            (ViewportDebugDraw)Enum.Parse(typeof(ViewportDebugDraw), mode, true)
+            Enum.Parse<ViewportDebugDraw>(mode, true)
         );
         }
 
@@ -66,7 +64,7 @@ public sealed class View : ICommand
     {
         ViewportSetDebugDraw(_yat.GetViewport().GetViewportRid(), debugDraw);
 
-        return ICommand.Success($"Set viewport debug draw to {debugDraw} ({(int)debugDraw}).");
+        return ICommand.Success(message: $"Set viewport debug draw to {debugDraw} ({(int)debugDraw}).");
     }
 
     private static bool IsValidMode(int mode)