+++ addons/yat/src/commands/Whereami.cs
@@ -12,8 +12,6 @@ namespace YAT.Commands;
 [Option("-e", "bool", "Prints the path to the executable.")]
 public sealed class Whereami : ICommand
 {
-    public string[]? Arguments { get; set; }
-
     public CommandResult Execute(CommandData data)
     {
         var s = (bool)data.Options["-s"];
@@ -21,7 +19,7 @@ public sealed class Whereami : ICommand
         var longForm = (bool)data.Options["-l"];
 
         return ICommand.Ok(
-            e
+            message: e
             ? GetExecutablePath(data.Terminal)
             : GetSceneInfo(data.Terminal, s, longForm)
         );