+++ addons/yat/src/types/CommandResult.cs
@@ -2,4 +2,8 @@ using YAT.Enums;
 
 namespace YAT.Types;
 
-public record CommandResult(ECommandResult Result, string Message = "");
+public record CommandResult(
+    ECommandResult Result,
+    string[]? OutData = null, // TODO: Allow OutData to be an any object.
+    string Message = ""
+);